bgpd, lib, ripnd: Isolate aggreagate pointer to using protocol
The `void *aggregate` pointer in ROUTE_NODE_FIELDS is only used by ripngd
and the bgp rfapi code.  As such every single route_node in the system is
paying the cost of keeping this pointer around and never using it.
Follow the table.h design pattern in bgp_table.h and create a
`struct ripng_node` and `struct rfapi_node` pointers that actually have
this aggregate pointer and then provide rfapi and ripng specific
struct route_node <-> struct rfapi/ripng_node translator functions.

On a 64 bit system this should account for 8 bytes of data per route_node
which is not insignificant considering a full bgp feed in both bgp and
zebra, or evpn routes which make signficant usage of tables.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>