lib: Add stream last pointer
Add a `static struct stream *last` pointer to stream.c.

This code will store the last stream sent to stream_free,
if the size of the stream is bigger than the last pointer
switch it out for that one.

On stream_new call if last exists and the size constraint
is satisfied then pass back the last pointer, else create
a new one.

In testing of 10k stream size and 10000000 allocations/frees
we reduce time from .65 to .06 seconds.

I am making an assumption here that there are two types
of stream allocations:

1) Initial setup of size X and we hold onto this pointer
until the end of the program.  In this case we might
have a couple of suboptimal( oversized ) stream allocations
but I think this is ok considering that the largest stream
allocation I am aware of is 65k.

2) Temporary scratch pads for the passing of data between
processes( zapi comes to mind immediately ), in which
case the stream data structure is ephermeral and constantly
alloced/freed.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>