zebra: handle renamed interfaces
when a RTM_NEWLINK netlink message is received, check if the interface
index is not already available among the already created interfaces in
the local zebra namespace. If present, and if the name of the interface
needs to be updated, then do it.
Also, opens the possibility to send again interface_add to daemon
client, so that it becomes possible to update the interface name in the
daemon interface list.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>