ospfd : issue with the "default-information originate always" command #2980
These changes are involves to resolve two issues

Issue-1 :  As reported ,Issue appears when user configures “default-information originate always “ command
preceded with negate command .

Root cause : While giving this config , type-5 lsa will be generated and flooded in AS irrespective of
presence of this default route in its RIB. While removing this config , this lsa will be assigned with
max age and schedule to remove  after max age delay timer (60 secs) expiry.
Before this timer expiryy , if user tried to configure again , it will fail due to subsequent checks in
the flow since the same lsa already present in the lsa database.But second time configuration  will be
succeeded before this timer expiry , since it is taking a different flow and flushing the existing lsa
and regenerating new lsa.

As per the current design , A message will be send to Zebra to validate the default route existence in
the RIB for originating.This may not be require if user provided “always” keyword I.e default route should
be originate and flood irrespective of its existence in the RIB.So this has been addressed now such that
message send to zebra only when user doesn’t provide “always” keyword.And the actual issue is addressed
by replacing the current mechanism with Refresh api . This ensures that generating new lsa if it is not
available or flushing and generating lsa if it is already available .

Issue -2 : similar issue observed with “default-information originate “ command I.e with out always keyword.
This fails due to a different reason.
Root cause :   when user configured this command , A message will be sent to zebra to validate default
route existence if available zebra send back message to ospf .
As part of processing this message , ospf will create a external list with type as “ZEBRA_KERNEL_ROUTE”
and update this route information in this external node. But in deletion flow , this external node is
being freed based on “DEFAULT_ROUTE” type . Due to this inconsistent type updation , it fails to delete
this external Node .
Which intern  make it fails for next subsequent “default-information originate” configuration .
To address this , made this type is consistent is both deletion and addition flow.

Signed-off-by: rgirada < rgirada@vmware.com >