bgpd: allow nht registration on ipv6 link-local addresses
Signed-off-by: Don Slice <dslice@cumulusnetworks.com>
Testing Done: bgp-smoke completed with no new failures

While testing 5549 support using global addresses, discovered that
ipv6 nexthop tracking thru a route-reflector didn't work.  Since
the next-hop used for remote nexthops resolves to the link-local
of the route-reflector, we need to track it in order to react to
interface down events.  Also tripped over a crash in certain cases
which is also resolved in this fix.