doc: Use `mv -f` in Makefile
Sphinx always runs, even in the `make install` stage. When `make install`
is run as root and then another `make` is run by a nonprivileged user,
some versions of `mv` prompt like this:

    mv: replace 'doc/manpages/_build/man/man.stamp',
                 overriding mode 0644 (rw-r--r--)?

Add -f to `mv` to avoid this. As `-f` is part of Posix, this should be
portable enough.

Signed-off-by: Christian Franke <chris@opensourcerouting.org>