lib: avoid creating interface when iface with previous index exists
On some cases, a zebra interface add event is sent to inform the daemon
that the interface name changed. In that case, the list of interface of
daemon is refreshed.
Instead of searching by interface name, a primary search by ifindex is
triggered.
As the vrf interface list by index is not maintained, that list is
populated.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>